[v0.9.8.3-types2] fix bug and rm several warnings. gamesnum_t set to 64 bits for now

+++ indiv.c
@@ -273,17 +273,17 @@ rate_super_players_internal
 					size_t n = myenc_n;
 					if (myenc[n].wh == j) { 
 						int opp = myenc[n].bl;
-						weig[r	] = myenc[n].played;
+						weig[r	] = (double)myenc[n].played;
 						rtng[r++] = ratingof[opp] - white_advantage;
 						cume_score += myenc[n].wscore;
-						cume_total += myenc[n].played;
+						cume_total += (double)myenc[n].played;
 				 	} else 
 					if (myenc[myenc_n].bl == j) { 
 						int opp = myenc[n].wh;
-						weig[r	] = myenc[n].played;
+						weig[r	] = (double)myenc[n].played;
 						rtng[r++] = ratingof[opp] + white_advantage;
-						cume_score += myenc[n].played - myenc[n].wscore;
-						cume_total += myenc[n].played;
+						cume_score += (double)myenc[n].played - myenc[n].wscore;
+						cume_total += (double)myenc[n].played;
 					} else {
 						fprintf(stderr,"ERROR!! function rate_super_players()\n");
 						exit(0);
